gfs2: gfs2_walk_metadata fix
It turns out that the current version of gfs2_metadata_walker suffers
from multiple problems that can lead to incorrect gfs2_hole_size
results.  To fix that, change gfs2_hole_walker to compute the metapath
to the first data block after the hole (if any), and compute the hole
size based on that.

Fixes fstest generic/490.

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>
Reviewed-by: Bob Peterson <rpeterso@redhat.com>