gfs2: Always mark inode dirty in fallocate
When allocating space with fallocate, always update the file timestamps
and mark the inode dirty, no matter if the FALLOC_FL_KEEP_SIZE flag set
or not.  The inode needs to be marked dirty so that a subsequent fsync
will pick up the inode, and any new allocations will make it to disk.
Filesystems like xfs and ext4 always update the timestamps, so make gfs2
behave the same way.

Fixes fstest generic/483.

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>