mm: memcontrol: flush percpu slab vmstats on kmem offlining
I've noticed that the "slab" value in memory.stat is sometimes 0, even if
some children memory cgroups have a non-zero "slab" value.  The following
investigation showed that this is the result of the kmem_cache reparenting
in combination with the per-cpu batching of slab vmstats.

At the offlining some vmstat value may leave in the percpu cache, not
being propagated upwards by the cgroup hierarchy.  It means that stats on
ancestor levels are lower than actual.  Later when slab pages are
released, the precise number of pages is substracted on the parent level,
making the value negative.  We don't show negative values, 0 is printed
instead.

To fix this issue, let's flush percpu slab memcg and lruvec stats on memcg
offlining.  This guarantees that numbers on all ancestor levels are
accurate and match the actual number of outstanding slab pages.

Link: http://lkml.kernel.org/r/20190819202338.363363-3-guro@fb.com
Fixes: fb2f2b0adb98 ("mm: memcg/slab: reparent memcg kmem_caches on cgroup removal")
Signed-off-by: Roman Gushchin <guro@fb.com>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Michal Hocko <mhocko@kernel.org>
Cc: Vladimir Davydov <vdavydov.dev@gmail.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>