FROMLIST: media: hantro: h264: Rename POC_CMP() into HANTRO_CMP()
And use it for all native type comparisons, even if it's not strictly
required. By doing that we make the code more consistent and prevent
from potential incorrect results in case of overflow or when the the
values being compared are both negative.

BUG=chromium:965375,chromium:1000144
TEST=emerge-veyron_minnie-kernelnext chromeos-kernel-4_19

Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>
(am from https://patchwork.kernel.org/patch/11137361)
Signed-off-by: Francois Buergisser <fbuergisser@chromium.org>

Change-Id: I03c55e29f2e7df6f7c947af7e3c550978fb1a6d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1797764
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Tested-by: Francois Buergisser <fbuergisser@chromium.org>
Commit-Queue: Francois Buergisser <fbuergisser@chromium.org>