FROMLIST: media: platform: mtk-mdp3: Add struct tuning_addr and img_sw_buffer
We added a struct tuning_addr which contains a field "present"
so that the driver can tell the firmware if we have user tuning
dataor not.

The strcut img_sw_buffer is also added. This struct has no cpu address
field and uses a handle instead so that we don't pass a cpu address
to co-processor

(am from https://patchwork.kernel.org/patch/11138515/)

TEST=build pass
BUG=b:140012892, b:111289971

Change-Id: I2f38d4045b2709cacbc2e4f5100a9a1c9f2d5c3f
Signed-off-by: Frederic Chen <frederic.chen@mediatek.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1798066
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>