FROMGIT: HID: google: whiskers: signal tablet mode on connect
When we receive "keyboard position" event from Whiskers we can
conclude that the base is attached, even if we did not get EC event
for that. We may miss EC event because there are some units which
have a lot of leakage on the ADC pins that the EC uses to determine
whether or not a base is attached.

Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit b543db46b47dc075535a5adf93460b611d669598
 git://git.kernel.org/pub/scm/linux/kernel/git/hid/hid.git for-5.5/whiskers)

BUG=b:140906956
TEST=Boot krane, attach masterball

Change-Id: I883203438e76f90f24e2e6c8ac5912aa6cf98434
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1804022