CHROMIUM: ASoC: max98090: Re-add shutdown events to ADC{L,R} controls.
This is a partial revert of commit b0fa9ee035b7 ("CHROMIUM: ASoC:
max98090: Enable DMIC at playback when selected.") - specifically, the
parts affecting the external mic.

This restores the audio output drop on cyan (matching 3.18) when
starting a recording with the external mic, in exchange for not
introducing noise on the external mic some of the time.

BUG=b:139918059
TEST=Recording with an external mic isn't noisy on cyan running 4.19.

Change-Id: Ia4e08069aeebbcf675cc72d1595f565cac39384c
Signed-off-by: Sam McNally <sammc@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1833372
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>