CHROMIUM: [media] hantro: fix motion vectors usage flag
The setting of the motion vectors usage flag and the setting of motion
vectors address do not match.
When decoding pre-recorded videos, this results of leaving the motion
vectors address unset.
In turn, the chip is accessing a non-existent address which will cause
iommu page faults.
This patch fixes the motion vector usage flag to match the condition
setting motion vectors address.

BUG=chromium:1008987
TEST=recorded video and played it back to confirm the green screen could
not been seen on veyron_minnie-kernelnext.
TEST=ran tast for DecodeAccelH264 and DecodeAccelVP8 on
veyron_minnie-kernelnext.

Signed-off-by: Francois Buergisser <fbuergisser@chromium.org>
Change-Id: I9838f2963a1a9025a38ef3f7e0a6cc0155c4494d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1835483
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>