mm,thp: recheck each page before collapsing file THP
In collapse_file(), after locking the page, it is necessary to recheck
that the page is up-to-date.  Add PageUptodate() check for both shmem THP
and file THP.

Current khugepaged should not try to collapse dirty file THP, because it
is limited to read only text.  Add a PageDirty check and warning for file
THP.  This is added after page_mapping() check, because if the page is
truncated, it might be dirty.

Link: http://lkml.kernel.org/r/20191018180345.4188310-1-songliubraving@fb.com
Fixes: 99cb0dbd47a1 ("mm,thp: add read-only THP support for (non-shmem) FS")
Signed-off-by: Song Liu <songliubraving@fb.com>
Cc: Kirill A. Shutemov <kirill.shutemov@linux.intel.com>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Hugh Dickins <hughd@google.com>
Cc: William Kucharski <william.kucharski@oracle.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>