CHROMIUM: bluetooth: Enable filter duplicated adv offload.
This change will reduce by an order of magnitude the number
of advertisements the system will need to process when an LE
device that distributed an IRK is paired but isn't connected.

This isn't a complete solution since the filter duplicate is
a best effort solution at best from the controller, so in very
noisy environements, it is possible that the controller will
be unable to filter duplicates and will send them all up to the
host.

BUG=b:143696105
TEST=verify that advertisements duplicates are filtered after
pairing a bluetooth LE mouse and disconnecting it.

Change-Id: I9f21d6f8ab2aeb00069f4509fd43ed2387c659bd
Signed-off-by: Alain Michaud <alainmichaud@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1894011
Reviewed-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Reviewed-by: Michael Sun <michaelfsun@google.com>
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>