mm: adjust shuffle code to allow for future coalescing
Patch series "mm / virtio: Provide support for unused page reporting", v13.

This series provides an asynchronous means of reporting unused guest pages
to a hypervisor so that the memory associated with those pages can be
dropped and reused by other processes and/or guests on the host.  Using
this it is possible to avoid unnecessary I/O to disk and greatly improve
performance in the case of memory overcommit on the host.

When enabled it will allocate a set of statistics to track the number of
reported pages.  When the nr_free for a given free_area is greater than
this by the high water mark we will schedule a worker to begin allocating
the non-reported memory and to provide it to the reporting interface via a
scatterlist.

Currently this is only in use by virtio-balloon however there is the hope
that at some point in the future other hypervisors might be able to make
use of it.  In the virtio-balloon/QEMU implementation the hypervisor is
currently using MADV_DONTNEED to indicate to the host kernel that the page
is currently unused.  It will be faulted back into the guest the next time
the page is accessed.

To track if a page is reported or not the Uptodate flag was repurposed and
used as a Reported flag for Buddy pages.  While we are processing the
pages in a given zone we have a set of pointers we track called
reported_boundary that is used to keep our processing time to a minimum.
Without these we would have to iterate through all of the reported pages
which would become a significant burden.  I measured as much as a 20%
performance degradation without using the boundary pointers.  In the event
of something like compaction needing to process the zone at the same time
it currently resorts to resetting the boundary if it is rearranging the
list.  However in the future it could choose to delay processing the zone
if a flag is set indicating that a zone is being actively processed.

Below are the results from various benchmarks.  I primarily focused on two
tests.  The first is the will-it-scale/page_fault2 test, and the other is
a modified version of will-it-scale/page_fault1 that was enabled to use
THP.  I did this as it allows for better visibility into different parts
of the memory subsystem.  The guest is running with 32G for RAM on one
node of a E5-2630 v3.

Test                page_fault1 (THP)     page_fault2
Baseline         1  1209281.00  +/-0.47%   411314.00  +/-0.42%
                16  8804587.33  +/-1.80%  3419453.00  +/-1.80%

Patches applied  1  1209369.67  +/-0.06%   412187.00  +/-0.10%
                16  8812606.33  +/-0.06%  3435339.33  +/-1.82%

Patches enabled  1  1209104.67  +/-0.11%   413067.67  +/-0.43%
 MADV disabled  16  8835481.67  +/-0.29%  3463485.67  +/-0.50%

Patches enabled  1  1210367.67  +/-0.58%   416962.00  +/-0.14%
                16  8433236.00  +/-0.58%  3437897.67  +/-0.34%

The results above are for a baseline with a linux-next-20191031 kernel,
that kernel with this patch set applied but page reporting disabled in
virtio-balloon, patches applied but the madvise disabled by direct
assigning a device, and the patches applied and page reporting fully
enabled.  These results include the deviation seen between the average
value reported here versus the high and/or low value.  I observed that
during the test the memory usage for the first three tests never dropped
whereas with the patches fully enabled the VM would drop to using only a
few GB of the host's memory when switching from memhog to page fault
tests.

Most of the overhead seen with this patch set fully enabled is due to the
fact that accessing the reported pages will cause a page fault and the
host will have to zero the page before giving it back to the guest.  The
overall guest size is kept fairly small to only a few GB while the test is
running.  This overhead is much more visible when using THP than with
standard 4K pages.  As such for the case where the host memory is not
oversubscribed this results in a performance regression, however if the
host memory were oversubscribed this patch set should result in a
performance improvement as swapping memory from the host can be avoided.

A brief history on the background of unused page reporting can be found at:
https://lore.kernel.org/lkml/29f43d5796feed0dec8e8bb98b187d9dac03b900.camel@linux.intel.com/

This patch (of 6):

Move the head/tail adding logic out of the shuffle code and into the
__free_one_page function since ultimately that is where it is really
needed anyway.  By doing this we should be able to reduce the overhead and
can consolidate all of the list addition bits in one spot.

Link: http://lkml.kernel.org/r/20191105220207.15144.8826.stgit@localhost.localdomain
Signed-off-by: Alexander Duyck <alexander.h.duyck@linux.intel.com>
Acked-by: David Hildenbrand <david@redhat.com>
Reviewed-by: Dan Williams <dan.j.williams@intel.com>
Cc: Michael S. Tsirkin <mst@redhat.com>
Cc: Matthew Wilcox <willy@infradead.org>
Cc: Michal Hocko <mhocko@kernel.org>
Cc: Mel Gorman <mgorman@techsingularity.net>
Cc: Vlastimil Babka <vbabka@suse.cz>
Cc: Yang Zhang <yang.zhang.wz@gmail.com>
Cc: <konrad.wilk@oracle.com>
Cc: Pankaj Gupta <pagupta@redhat.com>
Cc: Rik van Riel <riel@surriel.com>
Cc: Luiz Capitulino <lcapitulino@redhat.com>
Cc: Dave Hansen <dave.hansen@intel.com>
Cc: Wei Wang <wei.w.wang@intel.com>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: Paolo Bonzini <pbonzini@redhat.com>
Cc: Dan Williams <dan.j.williams@intel.com>
Cc: Oscar Salvador <osalvador@suse.de>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>