lib: optimize cpumask_local_spread()
In multi-processor and NUMA system, I/O driver will find cpu cores that
which shall be bound IRQ.  When cpu cores in the local numa have been
used, it is better to find the node closest to the local numa node,
instead of choosing any online cpu immediately.

On Huawei Kunpeng 920 server, there are 4 NUMA node(0 -3) in the 2-cpu
system(0 - 1).  We perform PS (parameter server) business test, the
behavior of the service is that the client initiates a request through the
network card, the server responds to the request after calculation.  When
two PS processes run on node2 and node3 separately and the network card is
located on 'node2' which is in cpu1, the performance of node2 (26W QPS)
and node3 (22W QPS) was different.  It is better that the NIC queues are
bound to the cpu1 cores in turn, then XPS will also be properly
initialized, while cpumask_local_spread only considers the local node.
When the number of NIC queues exceeds the number of cores in the local
node, it returns to the online core directly.  So when PS runs on node3
sending a calculated request, the performance is not as good as the node2.
It is considered that the NIC and other I/O devices shall initialize the
interrupt binding, if the cores of the local node are used up, it is
reasonable to return the node closest to it.

Let's optimize it and find the nearest node through NUMA distance for the
non-local NUMA nodes.  The performance will be better if it return the
nearest node than the random node.

After this patch, the performance of the node3 is the same as node2 that
is 26W QPS when the network card is still in 'node2'.  Since it will
return the closest non-local NUMA code rather than random node, it is no
harm to others at least.

Link: http://lkml.kernel.org/r/1572863268-28585-1-git-send-email-zhangshaokun@hisilicon.com
Signed-off-by: yuqi jin <jinyuqi@huawei.com>
Signed-off-by: Shaokun Zhang <zhangshaokun@hisilicon.com>
Cc: Mike Rapoport <rppt@linux.ibm.com>
Cc: Paul Burton <paul.burton@mips.com>
Cc: Michal Hocko <mhocko@suse.com>
Cc: Michael Ellerman <mpe@ellerman.id.au>
Cc: Anshuman Khandual <anshuman.khandual@arm.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>